// src/enums/listing.enum.ts
export enum ListingStatus {
  PENDING_APPROVAL = 'PENDING_APPROVAL',
  PUBLISHED = 'PUBLISHED',
  REJECTED = 'REJECTED',
  UNAVAILABLE = 'UNAVAILABLE',
  ARCHIVED = 'ARCHIVED',
}

export enum RoomType {
  SINGLE_ROOM = 'SINGLE_ROOM',
  SHARED_ROOM = 'SHARED_ROOM',
  APARTMENT = 'APARTMENT',
  MINI_APARTMENT = 'MINI_APARTMENT',
  DORMITORY = 'DORMITORY',
}

export enum SortBy {
  NEWEST = 'NEWEST',
  LOWEST_PRICE = 'LOWEST_PRICE',
  HIGHEST_PRICE = 'HIGHEST_PRICE',
  LARGEST_AREA = 'LARGEST_AREA',
  MOST_RELEVANT = 'MOST_RELEVANT',
}